+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/black-forest-labs
 
+## 1.0.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies [53f3368]
+  - @ai-sdk/provider@3.0.0-beta.25
+  - @ai-sdk/provider-utils@4.0.0-beta.44
+
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [dce03c4]
+  - @ai-sdk/provider-utils@4.0.0-beta.43
+  - @ai-sdk/provider@3.0.0-beta.24
+
 ## 1.0.0-beta.13
 
 ### Patch Changes