+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ai-sdk/google-vertex
 
+## 4.0.0-beta.102
+
+### Patch Changes
+
+- Updated dependencies [53f3368]
+  - @ai-sdk/provider@3.0.0-beta.25
+  - @ai-sdk/anthropic@3.0.0-beta.76
+  - @ai-sdk/google@3.0.0-beta.66
+  - @ai-sdk/provider-utils@4.0.0-beta.44
+
+## 4.0.0-beta.101
+
+### Patch Changes
+
+- Updated dependencies [0ae783e]
+  - @ai-sdk/anthropic@3.0.0-beta.75
+
+## 4.0.0-beta.100
+
+### Patch Changes
+
+- Updated dependencies [dce03c4]
+  - @ai-sdk/provider-utils@4.0.0-beta.43
+  - @ai-sdk/anthropic@3.0.0-beta.74
+  - @ai-sdk/provider@3.0.0-beta.24
+  - @ai-sdk/google@3.0.0-beta.65
+
 ## 4.0.0-beta.99
 
 ### Patch Changes